+++ packages/canvas_core/test/op_builder_scene_text_test.dart
@@ -7,7 +7,6 @@ import 'package:canvas_core/src/render_plan/op_builder_scene.dart';
 import 'package:canvas_core/src/runtime/model/scene_document.dart';
 import 'package:canvas_core/src/services/services.dart';
 import 'package:canvas_core/src/services/services_context.dart';
-import 'package:canvas_core/src/services/text_measure.dart';
 import 'package:canvas_core/src/algorithms/layout/computed_scene.dart'
     show computeScene;
 import 'package:test/test.dart';
@@ -40,7 +39,6 @@ void main() {
     final services = CoreServices(
       tm: _FakeTextMeasurer(),
       images: _FakeImages(),
-      textMeasureCache: TextMeasureCache(),
     );
 
     const solidColor = 0xFF445566;
@@ -77,7 +75,6 @@ void main() {
     final services = CoreServices(
       tm: _FakeTextMeasurer(),
       images: _FakeImages(),
-      textMeasureCache: TextMeasureCache(),
     );
 
     const c1 = 0xFF112233;